requirements.txt
imagesize==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
numpy==1.14.2
opencv-python==3.4.4.19
pika==0.12.0
Pillow==5.3.0
pylint==1.8.4
pytesseract==0.2.5
redis==3.0.1
requests==2.18.4
scikit-image==0.14.1
scipy==1.1.0
tensorflow==1.10.0
urllib3==1.22